♻️refactor(column): 컬럼 생성, 수정, 삭제 전체적인 리펙토링
1
BoardService.findBoardById 매서드
private -> public 으로 변경

ADD
    - DeleteColumnRequestDto.java
    - EditColumnRequestDto.java

BREAKING CHANGE(중요한 변경 사항): 1

1  Closes #0  (닫을 이슈 번호)

+++ src/main/java/com/sparta/bitbucket/column/service/ColumnService.java
@@ -1,11 +1,17 @@
 package com.sparta.bitbucket.column.service;
 
+import java.util.Objects;
+
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.sparta.bitbucket.auth.entity.Role;
 import com.sparta.bitbucket.auth.entity.User;
+import com.sparta.bitbucket.board.entity.Board;
+import com.sparta.bitbucket.board.service.BoardService;
 import com.sparta.bitbucket.column.dto.CreateColumnRequestDto;
+import com.sparta.bitbucket.column.dto.DeleteColumnRequestDto;
+import com.sparta.bitbucket.column.dto.EditColumnRequestDto;
 import com.sparta.bitbucket.column.entity.Columns;
 import com.sparta.bitbucket.column.repository.ColumnRepository;
 
@@ -16,12 +22,12 @@ import lombok.RequiredArgsConstructor;
 public class ColumnService {
 
 	private final ColumnRepository columnRepository;
-	private final BoardRepository boardRepository;
+	private final BoardService boardService;
 
 	@Transactional
-	public void createColumn(Long boardId, CreateColumnRequestDto requestDto, User user) {
-		Board board = validateUserAndBoard(boardId, user);
-		checkColumnTitleExists(boardId, requestDto.getTitle());
+	public void createColumn( User user, CreateColumnRequestDto requestDto) {
+		validateUser(user);
+		Board board = boardService.findBoardById(requestDto.getBoardId());
 
 		columnRepository.save(Columns.builder()
 			.title(requestDto.getTitle())
@@ -31,30 +37,38 @@ public class ColumnService {
 	}
 
 	@Transactional
-	public void deleteColumn(Long boardId, Long columnId, User user) {
-		validateUserAndBoard(boardId, user);
-
+	public void deleteColumn(Long columnId, User user, DeleteColumnRequestDto requestDto) {
+		validateUser(user);
+		checkUserIsBoardOwner(user, requestDto.getBoardId());
 		Columns byColumnId = findByColumnId(columnId);
 		columnRepository.delete(byColumnId);
 	}
 
-	public Board validateUserAndBoard(Long boardId, User user) {
+	@Transactional
+	public void updateColumn(Long columnId, User user, EditColumnRequestDto requestDto) {
+
+		validateUser(user);
+		checkUserIsBoardOwner(user, requestDto.getBoardId());
+		Columns columns = findByColumnId(columnId);
+
+		columns.toBuilder()
+			.title(requestDto.getTitle())
+			.orders(requestDto.getOrders())
+			.build();
+
+		columnRepository.save(columns);
+	}
+
+	public void validateUser(User user) {
 		if (user.getRole() != Role.MANAGER) {
 			throw new IllegalArgumentException("유저 권한이 없습니다."); // CommonException 로 바꿀예정
 		}
-
-		Board board = boardRepository.findById(boardId);
-
-		if (board.getUser.getId() != user.getId()) {
-			throw new IllegalArgumentException("유저아이디가 다릅니다."); // CommonException 로 바꿀예정
-		}
-
-		return board;
 	}
 
-	public void checkColumnTitleExists(Long boardId, String title) {
-		if (columnRepository.existsByBoardIdAndTitle(boardId, title)) {
-			throw new IllegalArgumentException("존재하는 컬럼 아이디입니다."); // CommonException 로 바꿀예정
+	public void checkUserIsBoardOwner(User user, Long boardId) {
+		Board board = boardService.findBoardById(boardId);
+		if(!Objects.equals(board.getUser().getId(), user.getId())) {
+			throw new IllegalArgumentException("유저 권한이 없습니다."); // CommonException 로 바꿀예정
 		}
 	}
 
@@ -63,4 +77,5 @@ public class ColumnService {
 			() -> new IllegalArgumentException("존재하지 않는 컬럼입니다.") // CommonException 로 바꿀예정
 		);
 	}
+
 }
